+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 
 - New work commands to make easier non technical users to use Hardis CI
   - **hardis:work:task:new**
-  - **hardis:work:task:save** 
+  - **hardis:work:task:save**
   - **hardis:work:task:refresh**
 
 ## [1.3.6] 2021-02-26